requirements.txt
python-telegram-bot
openai
requests
python-dotenv
notion-client
notion
asyncio
APScheduler
httpx
pytest
